[libmenu] Add a cache for listing the desktop files
When processing a layout, we get the list of desktop files from a set of
directories. It turns out that we nearly always use the same set of
directories, so adding a one-entry cache makes it possible to avoid
computing things again and again.

I changed Michael's patch a bit, mainly to empty the cache when the last
GMenuTree is unref'ed.

Closes: bgo#498749