[libmenu] Sort inlined items unless inline_header is used
We also strip duplicate entries in a menu when we do such a sort.

This is actually not a trivial task since it means we have to correctly
preprocess all inline data before using the layout -- else, we'd try to
sort items while a layout has been used, which is too late (since you
can specify that you want mixed entries and directories, eg; or even
entries or directories in a specified order).

We make sure to evaluate the inline data at the deepest first, so we can
correctly propagate as much inline items as possible to the top.

http://bugzilla.gnome.org/show_bug.cgi?id=490483