[libmenu,python] Add gmenu_tree_get_sort_key()/gmenu_tree_set_sort_key()
This is needed for users of gnome-menus that display applications by
strings obtained via get_display_name(), since the order of entry
changes if strings are different.

The default is to keep the sort order by name. At the moment, the only
other option is to use a sort by display name.

Bindings for python are added too.