[libmenu] Add real support for inline aliases during layout processing
It looks like the inline aliases were not really tested before...

All the processing of the layout info (merge_subdir_by_name,
merge_entry_by_id, merge_subdirs, merge_entries,
merge_subdirs_and_entries) was completely ignoring the fact that some
subdirs or entries might be aliases, and therefore need a different
processing.

This is now done correctly, and we have a merge_alias() function to do
the right thing when an alias is detected.

Also, this simplifies quite a bit the sorting of subdirs or entries
since we always need to check the types of the items we compare, so we
can get rid of gmenu_tree_directory_compare and
gmenu_tree_entry_compare.