libmenu: Clear cache of desktop entries set when files are added/removed
When installing or removing an application, we get a notification of
change. However, the menu tree still contained the same content.

We were simply not clearing the cache containing the list of desktop
entries when a file got added or removed, meaning that we always got the
same list of desktop entries.

(Note that it didn't affect gnome-panel, it's unclear why)

https://bugzilla.gnome.org/show_bug.cgi?id=630410