libmenu: Do not send multiple notifications for one file change
We emit notifications in the idle loop, which enables us to compress
multiple notifications: this way, we can check if there is already a
notification for a directory.

We do this at the CachedDir level because it enables us to get one
notification for a file rename, for example. But we also need it at the
MenuLayout level to be sure to compress notifications accross multiple
directories. It could be argued that we only need the latter, but I like
it this way :-)

https://bugzilla.gnome.org/show_bug.cgi?id=172046