layout: Show administration tools and old capplets in Other
Right now, administration tools and capplets that used to be in the
Control Center are not accessible in any way. With this patch, they
appear in Other. While this is not perfect, this will make them
accessible until we find the right solution.

To do this, we just exclude .desktop files that match panels from the
Control Center instead of excludings the Settings category, which is too
broad.

https://bugzilla.gnome.org/show_bug.cgi?id=645061