DesktopEntry: Clean up structure, make TryExec evaluation lazy
This is preparatory work for rebasing DesktopEntry on top of
GDesktopAppInfo.

First, it doesn't make sense to represet a random subset of booleans
as flags; just flatten these into a bitfield.  Move the refcount to be
a plain guint at top.

Second, previously we were calling g_find_program_in_path when
creating a .desktop file, even if the caller wasn't interested in
whether the TryExec succeeded.  Make this lazy.

https://bugzilla.gnome.org/show_bug.cgi?id=647968