Drop GMenuTree caching and support for absolute paths
This is work towards converting to GObject; the API to create a tree
is now just gmenu_tree_new().

First, the internal caching makes things very complex for little gain
- gnome-shell only creates one GMenuTree, and gnome-panel could pretty
easily be converted to do so.

In a Google Code Search, I couldn't find anyone using absolute paths
for menus, and looking through the revision history, I don't see
a rationale for it.  So, just drop it too.

https://bugzilla.gnome.org/show_bug.cgi?id=647968