libmenu: Add GMENU_TREE_FLAGS_INCLUDE_UNALLOCATED flag
Add a new GMENU_TREE_FLAGS_INCLUDE_UNALLOCATED flag to add in the root
directory entries that are not allocated anywhere else. This is useful
if the user really wants to get absolutely all entries (in addition to
using INCLUDE_EXCLUDED, which is a bit different, and
INCLUDE_NODISPLAY).

Add gmenu_tree_entry_get_is_unallocated() API matching this flag.

https://bugzilla.gnome.org/show_bug.cgi?id=668512