layout: Deal with some vendor renames
The menu spec doesn't offer very specific guidance for how
to use vendor prefixes. And it isn't even clear about what
a vendor is. Or how it helps avoid collisions with a naming scheme
that isn't fully qualified in the first place.

Longer term we should move to something that uses fully qualified
names like the dbus spec. In the meantime we have to deal with
"aliases" for app names.

https://bugzilla.gnome.org/show_bug.cgi?id=694131