gmenu-tree: Properly implement <OnlyUnallocated />
The implementation here before was bad -- if two directories explicitly
included some menu entry and had <OnlyUnallocated />, they'd both get the
menu file, which is obviously wrong.

Fix this so that if two directories have <OnlyUnallocated />, the first
one in the menu file "wins", which as far as I can tell is the correct
semantics.