layout: don't use OnlyUnallocated for sections different than Other
We used to have OnlyUnallocated on all non-folder sections, as
a quick way to enforce the whitelist, but that broke the Other
section.
Instead, we need a whitelist in the two folders, and a blacklist
in each category the application would normally be in (accessories
or system tools usually).
Ideally, apps in the Utilities category would expose the special
X-GNOME-Utilities desktop category, which we now recognize, but
the whitelist is kept to avoid patching all of them.
For the sundry category, ideally it should not be there at all.

In the process, a number of .desktop files were removed, as they're
not provided by any packages (checked jhbuild and Fedora 19). They
are:
* caribou
* fedora-im-chooser
* fedora-abrt
* gnome-orca
* gnome-user-share-properties
* ibus
* ibus-daemon

Distributions that are changing .desktop file names should take
care of patching gnome-menus appropriately.

https://bugzilla.gnome.org/show_bug.cgi?id=706763